chore(refactor): move cspc code v1alpha1 package
This PR -

- Moves CSPC builder code to v1alpha1 package
- Moves localdevice controller and finalizer to v1alpha1 package
- Moves metac.config to v1alpha1 directory

Here is the part of sts using this you can create CSPC of v1alpha1
using cstorpoolauto
```yaml
containers:
- args
  - --logtostderr
  - --run-as-local
  - -v=5
  - --discovery-interval=40s
  - --cache-flush-interval=240s
  - --metac-config-path=/etc/config/metac/localdevice/v1alpha1/
  command:
  - /usr/bin/cstorpoolauto
  image: mayadataio/cstorpoolauto:<>
  name: cstorpoolauto
```
