+++ common/cstorpoolcluster/v1alpha1/helper.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package cstorpoolcluster
+package v1alpha1
 
 import (
 	"github.com/pkg/errors"
