+++ controller/localdevice/v1alpha1/finalizer.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package localdevice
+package v1alpha1
 
 import (
 	"fmt"
