+++ controller/localdevice/v1alpha1/finalizer_test.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package localdevice
+package v1alpha1
 
 import (
 	"testing"
