+++ controller/localdevice/v1alpha1/reconciler.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package localdevice
+package v1alpha1
 
 import (
 	"fmt"
@@ -27,7 +27,7 @@ import (
 
 	bd "mayadata.io/cstorpoolauto/common/blockdevice"
 	ccc "mayadata.io/cstorpoolauto/common/cstorclusterconfig"
-	cspc "mayadata.io/cstorpoolauto/common/cstorpoolcluster"
+	cspc "mayadata.io/cstorpoolauto/common/cstorpoolcluster/v1alpha1"
 	metaccommon "mayadata.io/cstorpoolauto/common/metac"
 	"mayadata.io/cstorpoolauto/types"
 	"mayadata.io/cstorpoolauto/unstruct"
